refactor: table schema
